+++ internal/settings/settings_test.go
@@ -35,6 +35,7 @@ var (
 	anotherConfigPath    = filepath.Join(".", "testdata", "another-config.yaml")
 	localGatewayConfig   = filepath.Join("..", "..", "cmd/otelcol/config/collector/gateway_config.yaml")
 	localOTLPLinuxConfig = filepath.Join("..", "..", "cmd/otelcol/config/collector/otlp_config_linux.yaml")
+	propertiesPath       = filepath.Join(".", "testdata", "properties.yaml")
 )
 
 func TestNewSettingsWithUnknownFlagsNotAcceptable(t *testing.T) {
@@ -98,12 +99,16 @@ func TestNewSettingsConfMapProviders(t *testing.T) {
 	require.Contains(t, confMapProviders, settings.discovery.DiscoveryModeScheme())
 	discoveryModeProvider := confMapProviders[settings.discovery.DiscoveryModeScheme()]
 
+	require.Contains(t, confMapProviders, settings.discovery.PropertiesFileScheme())
+	propertiesFileProvider := confMapProviders[settings.discovery.PropertiesFileScheme()]
+
 	require.Equal(t, map[string]confmap.Provider{
-		envProvider.Scheme():                     envProvider,
-		fileProvider.Scheme():                    fileProvider,
-		settings.discovery.PropertyScheme():      propertyProvider,
-		settings.discovery.ConfigDScheme():       configdProvider,
-		settings.discovery.DiscoveryModeScheme(): discoveryModeProvider,
+		envProvider.Scheme():                      envProvider,
+		fileProvider.Scheme():                     fileProvider,
+		settings.discovery.PropertyScheme():       propertyProvider,
+		settings.discovery.ConfigDScheme():        configdProvider,
+		settings.discovery.DiscoveryModeScheme():  discoveryModeProvider,
+		settings.discovery.PropertiesFileScheme(): propertiesFileProvider,
 	}, confMapProviders)
 }
 
@@ -524,6 +529,7 @@ func TestDefaultDiscoveryConfigDir(t *testing.T) {
 	require.NoError(t, err)
 	require.True(t, settings.discoveryMode)
 	require.False(t, settings.configD)
+	require.Nil(t, settings.discoveryPropertiesFile.value)
 
 	require.Equal(t, []string{
 		localGatewayConfig,
@@ -546,16 +552,29 @@ func TestInheritedDiscoveryConfigDir(t *testing.T) {
 
 func TestInheritedDiscoveryConfigDirWithConfigD(t *testing.T) {
 	t.Cleanup(setRequiredEnvVars(t))
-	settings, err := New([]string{"--discovery", "--config-dir", "/some/config.d", "--configd"})
+	settings, err := New([]string{
+		"--discovery", "--config-dir", "/some/config.d", "--configd", "--discovery-properties", propertiesPath,
+	})
 	require.NoError(t, err)
 	require.True(t, settings.discoveryMode)
 	require.True(t, settings.configD)
 
+	require.NotNil(t, settings.discoveryPropertiesFile.value)
+	require.Equal(t, propertiesPath, settings.discoveryPropertiesFile.String())
+
 	require.Equal(t, []string{
 		localGatewayConfig,
+		fmt.Sprintf("splunk.properties:%s", propertiesPath),
 		"splunk.configd:/some/config.d",
 		"splunk.discovery:/some/config.d",
 	}, settings.ResolverURIs())
+}
+
+func TestDiscoveryPropertiesMustExist(t *testing.T) {
+	t.Cleanup(setRequiredEnvVars(t))
+	settings, err := New([]string{"--discovery", "--discovery-properties", "notafile"})
+	require.ErrorContains(t, err, "unable to find discovery properties file notafile. Ensure flag '--discovery-properties' is set correctly:")
+	require.Nil(t, settings)
 }
 
 // to satisfy Settings generation